server/requirements.txt
fastmcp==3.2.4
httpx>=0.28.1
psycopg2-binary>=2.9.9
